requirements.txt
flask==2.3.3
pypdf==4.0.1
langchain==0.1.5
langchain-google-genai==0.0.5
langchain-core==0.1.15
langgraph==0.0.23
murf==1.7.1
python-dotenv==1.0.0
werkzeug==2.3.7
gunicorn==21.2.0